+++ src/components/Fieldsets/RegistrationEmailFieldset.js
@@ -11,15 +11,17 @@ const isProduction = !$IS_DEVELOPMENT && !$IS_STAGING
 
 const RegistrationEmailFieldset = forwardRef((props, ref) => {
   const {
+    children,
     onValidate: parentValidate,
     ...inputProps
   } = props
 
   const handleValidate = useValidationCallback(
     (messages, value) => {
-      if (value === 'surly_badger@fuelrats.com') {
+      const compareValue = value.toLowerCase().trim()
+      if (compareValue === 'surly_badger@fuelrats.com') {
         messages.errors.push('While we appreciate your admiration for our founder, this email is most surely not your own.')
-      } else if (isProduction && value.includes('@fuelrats')) {
+      } else if (isProduction && compareValue.includes('@fuelrats')) {
         messages.errors.push('Fuel rats email accounts cannot be used to register a brand new account, goofball.')
       }
     },
@@ -31,7 +33,9 @@ const RegistrationEmailFieldset = forwardRef((props, ref) => {
     <EmailFieldset
       ref={ref}
       {...inputProps}
-      onValidate={handleValidate} />
+      onValidate={handleValidate}>
+      {children}
+    </EmailFieldset>
   )
 })
 RegistrationEmailFieldset.displayName = 'RegistrationEmailFieldset'